Rework packages, install, and discovery to run

+++ tools/discovery-load.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+set -e
+
+if [ ! -e assets ] ; then
+        echo "No assets directory to work from."
+        exit 1
+fi
+
+cd assets
+
 export RS_KEY=${RS_KEY:-rocketskates:r0cketsk8ts}
 ./drpcli bootenvs install bootenvs/local.yml
 ./drpcli bootenvs install bootenvs/discovery.yml
